refactor(#3847): campsite-split frozen map decoding out of frozen.rs (file-size ratchet)
The gate of record at `f30837441` returned RESULT: FAIL on ONE component —
`file-size`, the campsite ratchet — with tree-integrity PASS and 38/38
component-set. 37 of 38 components were otherwise fine, so the three decoder
fixes held up under the full gate; this commit is about where the code LIVES.

What the ratchet caught, and it was right:

    cell_path_key.rs        792 -> 830   (limit 800)  I pushed it OVER
    cell_path_key_tests.rs 1494 -> 1531  (limit 1500) I pushed it OVER
    frozen.rs               933 -> 1017  (limit 800)  already over; I grew it

Two of three went over from UNDER, and a large share of the growth was my own
explanatory comment prose.

WHY NOT `CQLITE_ALLOW_FILE_GROWTH=1`. It is documented and legitimate, but a
re-gate was unavoidable either way once the component FAILed, so the override
would buy nothing except a permanent OPT-OUT row on the record. And a split was
NOT out of scope: the growth was my own additions, not pre-existing bulk.
`frozen.rs` was 133 lines over threshold BEFORE #3847 went near it, which means
the ratchet forbids ANY growth there and two call sites cannot be zero lines —
that is precisely the case the campsite rule exists for.

The split, by responsibility rather than by line count:

  * NEW `row_decoder/frozen_map.rs` (280) — frozen MAP decoding
    (`parse_frozen_map_value`, `parse_frozen_map_value_raw`) plus the never-null
    map-key rule they now enforce, and the header carrying the whole #3847
    argument: one root cause, four defects, and the three properties that stop a
    fifth site appearing. `frozen.rs` keeps list/set/tuple.
  * NEW `complex_column/cell_path_key_tests_frozen.rs` (49) — a campsite split of
    a tests file sitting exactly at its 1500-line threshold (#1135). It lives
    under `complex_column` and NOT beside the rest of the #3847 key work because
    `parse_cell_path_key` is `pub(super)` there: a sibling module cannot call it.
    The header says so, so nobody consolidates it back and rediscovers that.
  * `frozen.rs` 1017 -> 795 — UNDER its 800 threshold for the first time.

Two edits also SHRANK the diff, because my originals were simply worse code: I
had replaced a pre-existing 7-line `Err(_) if …` guard arm with a 10-line
`Err(e) => { if … }` block doing the same thing, and `peeled_for_inspection`'s
`pub(in crate::…::row_decoder)` forced a three-line wrap where `pub(crate)` is one
line for an internal helper. `read_frozen_element` becomes `pub(super)` so the
moved decoders can reach it.

Full `cqlite-core` lib suite unchanged at 3832 passed / 0 failed under
`-D warnings` — same count as before the split, which is the point.

Refs #3847
