+++ cqlite-core/src/storage/sstable/reader/parsing/row_decoder/complex_column.rs
@@ -11,6 +11,8 @@ mod cell_path_key;
 #[cfg(test)]
 mod cell_path_key_tests;
 #[cfg(test)]
+mod cell_path_key_tests_frozen; // campsite split of cell_path_key_tests (#1135)
+#[cfg(test)]
 mod regression_3747_empty_map_key_tests;
 // Issue #3612 (R3-F1): the guarded component-length conversion, shared with the
 // UDT field loops in `udt.rs` / `raw_type_value.rs` (see that module's header for
